+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 12.6.0-beta.1 (Unreleased)
-
+- Fixed bug where DataLakeServiceClient.GetFileSystemClient(), DataLakeFileSystemClient.GetFileClient(), DataLakeFileSystemClient.GetDirectoryClient(), DataLakeDirectoryClient.GetSubDirectoryClient() and DataLakeFileClient.GetFileClient() created clients that could not generate a SAS from clients that could generate a SAS
 
 ## 12.5.0 (2020-11-10)
 - Includes all features from 12.5.0-preview.1